requirements.txt
Django==4.2.23
psycopg2-binary==2.9.10
Pillow==11.2.1
requests==2.32.4
python-decouple==3.8
django-cors-headers==4.7.0
reportlab==4.2.5
matplotlib==3.8.4
seaborn==0.13.2
pandas==2.2.3
openpyxl==3.1.5
celery==5.3.7
redis==5.0.8
